separate functions into modules and add debug toml
